Restructure app.js, routes and controllers
