+++ .gitignore
@@ -131,3 +131,5 @@ dist
 
 # ignore package-lock.json
 package-lock.json
+
+.DS_Store
